+++ src/sdk/attributes.ts
@@ -98,7 +98,7 @@ export class Attributes {
           if (utils.matchContentType(contentType, `application/json`)) {
             res.attributeGroupResponse = [];
             const resFieldDepth: number = utils.getResFieldDepth(res);
-            res.attributeGroupResponse = utils.deserializeJSONResponse(
+            res.attributeGroupResponse = utils.objectToClass(
               httpRes?.data,
               shared.AttributeGroupResponse,
               resFieldDepth
@@ -107,7 +107,7 @@ export class Attributes {
           break;
         case httpRes?.status == 400:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.clientError = utils.deserializeJSONResponse(
+            res.clientError = utils.objectToClass(
               httpRes?.data,
               shared.ClientError
             );
@@ -115,7 +115,7 @@ export class Attributes {
           break;
         case httpRes?.status == 500:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.serverError = utils.deserializeJSONResponse(
+            res.serverError = utils.objectToClass(
               httpRes?.data,
               shared.ServerError
             );
@@ -189,7 +189,7 @@ export class Attributes {
       switch (true) {
         case httpRes?.status == 200:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.attributeMappingResponse = utils.deserializeJSONResponse(
+            res.attributeMappingResponse = utils.objectToClass(
               httpRes?.data,
               shared.AttributeMappingResponse
             );
@@ -197,7 +197,7 @@ export class Attributes {
           break;
         case httpRes?.status == 400:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.clientError = utils.deserializeJSONResponse(
+            res.clientError = utils.objectToClass(
               httpRes?.data,
               shared.ClientError
             );
@@ -205,7 +205,7 @@ export class Attributes {
           break;
         case httpRes?.status == 500:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.serverError = utils.deserializeJSONResponse(
+            res.serverError = utils.objectToClass(
               httpRes?.data,
               shared.ServerError
             );
@@ -279,7 +279,7 @@ export class Attributes {
       switch (true) {
         case httpRes?.status == 200:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.bulkAttributeResponse = utils.deserializeJSONResponse(
+            res.bulkAttributeResponse = utils.objectToClass(
               httpRes?.data,
               shared.BulkAttributeResponse
             );
@@ -287,7 +287,7 @@ export class Attributes {
           break;
         case httpRes?.status == 400:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.clientError = utils.deserializeJSONResponse(
+            res.clientError = utils.objectToClass(
               httpRes?.data,
               shared.ClientError
             );
@@ -295,7 +295,7 @@ export class Attributes {
           break;
         case httpRes?.status == 500:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.serverError = utils.deserializeJSONResponse(
+            res.serverError = utils.objectToClass(
               httpRes?.data,
               shared.ServerError
             );
@@ -369,7 +369,7 @@ export class Attributes {
       switch (true) {
         case httpRes?.status == 200:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.bulkAttributeResponse = utils.deserializeJSONResponse(
+            res.bulkAttributeResponse = utils.objectToClass(
               httpRes?.data,
               shared.BulkAttributeResponse
             );
@@ -377,7 +377,7 @@ export class Attributes {
           break;
         case httpRes?.status == 400:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.clientError = utils.deserializeJSONResponse(
+            res.clientError = utils.objectToClass(
               httpRes?.data,
               shared.ClientError
             );
@@ -385,7 +385,7 @@ export class Attributes {
           break;
         case httpRes?.status == 500:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.serverError = utils.deserializeJSONResponse(
+            res.serverError = utils.objectToClass(
               httpRes?.data,
               shared.ServerError
             );
@@ -460,7 +460,7 @@ export class Attributes {
       switch (true) {
         case httpRes?.status == 200:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.attributeGroupSearchResponse = utils.deserializeJSONResponse(
+            res.attributeGroupSearchResponse = utils.objectToClass(
               httpRes?.data,
               shared.AttributeGroupSearchResponse
             );
@@ -468,7 +468,7 @@ export class Attributes {
           break;
         case httpRes?.status == 400:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.clientError = utils.deserializeJSONResponse(
+            res.clientError = utils.objectToClass(
               httpRes?.data,
               shared.ClientError
             );
@@ -476,7 +476,7 @@ export class Attributes {
           break;
         case httpRes?.status == 500:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.serverError = utils.deserializeJSONResponse(
+            res.serverError = utils.objectToClass(
               httpRes?.data,
               shared.ServerError
             );
@@ -551,16 +551,15 @@ export class Attributes {
       switch (true) {
         case httpRes?.status == 200:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.attributeMappingPaginationResponse =
-              utils.deserializeJSONResponse(
-                httpRes?.data,
-                shared.AttributeMappingPaginationResponse
-              );
+            res.attributeMappingPaginationResponse = utils.objectToClass(
+              httpRes?.data,
+              shared.AttributeMappingPaginationResponse
+            );
           }
           break;
         case httpRes?.status == 400:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.clientError = utils.deserializeJSONResponse(
+            res.clientError = utils.objectToClass(
               httpRes?.data,
               shared.ClientError
             );
@@ -568,7 +567,7 @@ export class Attributes {
           break;
         case httpRes?.status == 500:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.serverError = utils.deserializeJSONResponse(
+            res.serverError = utils.objectToClass(
               httpRes?.data,
               shared.ServerError
             );